requirements_win.txt
pywin32
moviepy
python-pptx
azure-cognitiveservices-speech
python-dotenv